+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+### 1.0.1
+- Changed all the 'PathfinderMobs' to be aggressive (it were only 'Animals' before)
+- Made the 'NeutralMobs' always be aggressive to the player
+
 ## 1.0.0
 - Initial release